+++ tests/cancel6d.c
@@ -165,11 +165,11 @@ main()
   for (i = 1; i <= NUMTHREADS; i++)
     {
       int fail = 0;
-      int result = 0;
+      intptr_t result = 0;
 
       assert(pthread_join(t[i], (void **) &result) == 0);
 
-      fail = (result != (int) (size_t) PTHREAD_CANCELED);
+      fail = (result != (intptr_t) PTHREAD_CANCELED);
 
       if (fail)
 	{
